fix bug tiempo

+++ src/asterix_lib/cat10.cpp
@@ -417,7 +417,7 @@ void Cat10:: DecodeTargetReportDescriptor(QVector<unsigned char> &dataItem) {
 void Cat10::DecodeTimeOfDay(QVector<unsigned char> &dataItem) {
     double timeResolution = pow(2,-7);
     double seconds = Utilities::DataTools::DecodeUnsignedBytesToDouble(dataItem,timeResolution);
-    int mseconds = (int) seconds * 1000;
+    int mseconds = seconds * 1000;
     this->timeOfDay = QTime::fromMSecsSinceStartOfDay(mseconds);
 }
