+++ src/asterix_lib/cat19.cpp
@@ -127,7 +127,7 @@ void Cat19::DecodeMessageType(QVector<unsigned char> &dataItem) {
 void Cat19::DecodeTimeOfDay(QVector<unsigned char> &dataItem) {
     double timeResolution = pow(2,-7);
     double seconds = Utilities::DataTools::DecodeUnsignedBytesToDouble(dataItem,timeResolution);
-    int mseconds = (int) seconds * 1000;
+    int mseconds = seconds * 1000;
     this->timeOfDay = QTime::fromMSecsSinceStartOfDay(mseconds);
 }
